+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Next major release v2, includes various new features and core changes. The API i
 * CORE: [!BREAKER] \Micro provides an OpenID-Connect authentication adapter, the current oauth2 auth adapter \Balloon\Auth\Adapter\Oauth2 gets removed with this release (see upgrade guide) #8
 * CORE: [CHANGE] changed hook preAuthentication() first param to Auth $auth instead auth adapters
 * CORE: [CHANGE] Moved various namespaces: \Balloon\Rest => \Balloon\Api, \Balloon\Plugin => \Balloon\Hook #55, \Balloon\Queue => \Balloon\Async
-* CORE: [CHANGE] renamed \Ballon\Exception\Coding to \Balloon\Exception\Internal
+* CORE: [CHANGE] PHP set_error_handler now throws ErrorException instead \Balloon\Exception\Coding
 * CORE: [CHANGE] Added new \Balloon\Server which is the new point of entry, also moved \Balloon\User to \Ballon\Server\User and made various code improvements to it
 * CORE: [CHANGE] Moved \Balloon\Filesystem\node\INode to \Balloon\Filesystem\node\NodeInterface and \Balloon\Filesystem\Node\Node to \Balloon\Filesystem\Node\AbstractNode #6
 * CORE: [CHANGE] Elasticsearch is now an app and not part of the core anymore #10
@@ -29,6 +29,8 @@ Next major release v2, includes various new features and core changes. The API i
 * CORE: [CHANGE] Changed generating access token to random_bytes() for creating sharelink tokens
 * CORE: [FEATURE] added a couple of new methods to NodeAbstract to set/receive/unset app based attributes for invidual nodes 
 * CORE: [CHANGE] added AbstractNode::getAttributes(array $attributes=[]) besides AbstractNode::getAttribute()
+* CORE: [FIX] fixed application/octet-stream mime type for office files (issue since 1.x)
+* CORE: [CHANGE] Extracted Mime detection to \Balloon\Mime 
 * API: [CHANGE] removed GET /api/v1/about
 * API: [CHANGE] removed GET /api/v1/version
 * API: [CHANGE] added 'name' to output of GET /api and GET /api/v1 #46